Add API hide and show toggle with eye icon (#773)
* feat: implement password field with toggle for API key visibility

* refactor: addressed feedbacks related to documentation
