+++ pages/godam/components/tabs/VideoSettings/APISettings.jsx
@@ -6,7 +6,7 @@ import { useState } from 'react';
 /**
  * WordPress dependencies
  */
-import { Button, Panel, PanelBody, TextControl, Spinner } from '@wordpress/components';
+import { Button, Panel, PanelBody, Spinner } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 
 /**
@@ -15,26 +15,12 @@ import { __ } from '@wordpress/i18n';
 import { useDeactivateAPIKeyMutation, useVerifyAPIKeyMutation } from '../../../redux/api/media-settings.js';
 import { hasValidAPIKey, maskedAPIKey, scrollToTop } from '../../../utils/index.js';
 import UsageData from './UsageData.jsx';
+import PasswordFieldWithToggle from './components/PasswordFieldWIthToggle/index.jsx';
 
 const APISettings = ( { setNotice } ) => {
 	const [ apiKey, setAPIKey ] = useState( hasValidAPIKey ? maskedAPIKey : '' );
 	const [ verifyAPIKey, { isLoading: isAPIKeyLoading } ] = useVerifyAPIKeyMutation();
 	const [ deactivateAPIKey, { isLoading: isDeactivateLoading } ] = useDeactivateAPIKeyMutation();
-
-	// Function to render help text based on API key validity
-	const renderHelpText = () => {
-		if ( ! hasValidAPIKey ) {
-			return (
-				<>
-					{ __( 'Your API key is required to access the features. You can get your active API key from your ', 'godam' ) }
-					<a href={ ( window.godamRestRoute?.apiBase ?? 'https://app.godam.io' ) + '/web/my-account?accTab=API' } target="_blank" rel="noopener noreferrer" className="text-blue-500 underline">
-						{ __( 'Account', 'godam' ) }
-					</a>.
-				</>
-			);
-		}
-		return null;
-	};
 
 	// Function to handle saving the API key
 	const handleSaveAPIKey = async () => {
@@ -95,14 +81,11 @@ const APISettings = ( { setNotice } ) => {
 		<Panel header={ __( 'API Settings', 'godam' ) } className="godam-panel">
 			<PanelBody initialOpen className="flex gap-8 flex-col sm:flex-row">
 				<div className="flex flex-col gap-2 b-4m">
-					<TextControl
-						label={ __( 'API Key', 'godam' ) }
-						value={ apiKey }
-						onChange={ setAPIKey }
-						help={ renderHelpText() }
-						placeholder={ __( 'Enter your API key here', 'godam' ) }
-						className={ `godam-input ${ ! hasValidAPIKey && maskedAPIKey ? 'invalid-api-key' : '' }` }
-						disabled={ hasValidAPIKey }
+					<PasswordFieldWithToggle
+						hasValidAPIKey={ hasValidAPIKey }
+						maskedAPIKey={ maskedAPIKey }
+						apiKey={ apiKey }
+						setAPIKey={ setAPIKey }
 					/>
 					<div className="flex gap-2">
 						<Button
